Move some files around and delete old date picker
