--- common/components/inputs/DateSelectionDropdown.tsx
@@ -1,1 +1,0 @@
-export const DateSelectionDropdown = () => {};
